backend/requirements.txt
# FinPilot FastAPI Backend - Production Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
httpx==0.25.2
alembic==1.12.1
python-multipart==0.0.6
python-cors==1.7.0
google-generativeai>=0.3.0
requests==2.31.0

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
